refactor: Move BottomScreenCommon from models to utils package
